Update PhysiCell_pathology.cpp
Made the legend function safer. It just makes a temporary Cell that isn't registered into all the data structures long enough to run the coloring function and then delete as it goes out of scope. 

Nice benefit is that initial plots no longer show the "dummy" cell in the cell count.