Fixing the fix...
My initial fix unsynched the rules_map and the rules members of Hypothesis_Ruleset. This meant that when add_signal is called, only the Hypothesis_Rule's pointed to by the rules_map were updated and not the Hypothesis_Rule's in the rules vector.

In this fix, the rules vector is now a vector of pointers to Hypothesis_Rules. See Lines 780 and 781 where both end up with the same (pointer) value stored.