Fix substrate label placement for jpeg conversion
For some reason I cannot comprehend, the svg->jpeg conversion does not like individual text elements having a rotation, so now the substrate colorbar label is placed in a group that is rotated.

For another reason I cannot comprehend, the ordering of the svg elements affects how this rotation happens. If I move the substrate colorbar label up, then it is placed in the right place upon conversion to an svg.