fix attack no detach
PR #340 put the correct check in the wrong place to prevent detachment when a cell is attacking another cell.
This fixes that by first moving the check into the spring attachments function as an attack produces a spring attachment.
Second, it moves the check into the detachment block rather than the attachment block.